% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{modify,data.table,modifier-method}
\alias{modify,data.table,modifier-method}
\title{Modify records in a data.table}
\usage{
\S4method{modify}{data.table,modifier}(dat, x, copy = NULL, sequential = TRUE, ...)
}
\arguments{
\item{dat}{\code{\link[=data.table]{data.table()}} object}

\item{x}{\code{dcmodify::modifier} object.}

\item{copy}{if \code{TRUE} modify copy of table.}

\item{sequential}{if \code{TRUE} (default), steps will be executed in sequence, so order matters.}

\item{...}{unused}
}
\description{
Modify records in a data.table using modification rules specified
in a modifier object.
}
\details{
This is a more efficient implementation then coercing the data.table to a
data.frame and use that implementation.
}
\examples{
library(data.table)

m <- modifier( if (age > 130) age = 130
             , income[age < 12] <- 0
             )

dat <- fread(text =
"age, income
 140,  300
  11, 2000
  25, 3000"
)

# modify a copy of the data
dat_m <- modify(dat, m, copy = TRUE)
print(dat_m)

# the data it self
setmodify(dat, m)
print(dat)
}
