% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_dt.R
\name{dump_dt}
\alias{dump_dt}
\title{Dump data.table modification script}
\usage{
dump_dt(x, name = "dat", file = stdout())
}
\arguments{
\item{x}{\code{dcmodify::modifier()} object with the modification rules}

\item{name}{\code{character} name of data.table to be used.}

\item{file}{where should the generated file be written?}
}
\description{
Dump data.table modification script
}
