% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\name{setmodify}
\alias{setmodify}
\title{modifies data.table in place}
\usage{
setmodify(dat, x, ...)
}
\arguments{
\item{dat}{\code{\link[=data.table]{data.table()}} object}

\item{x}{\code{dcmodify::modifier()} object.}

\item{...}{not used}
}
\description{
modifies data.table in place, alias for \code{\link[=modify]{modify()}} with \code{copy=TRUE} and
\code{sequential=TRUE}
}
